Fixed peculiar let scope evaluation
This was.. strange. I had a let declaration for the same variable name on the
if and else of a control statement. Before the update to vue, this caused no
problem. After, it seemed the control statement was being evaluated such that
both declarations (one on the if, the other the else) were called. I do not know
why this is the case. Furthermore, vue created a property on the local scope that
was the variable name prefixed by an underscore (_dequeue) to hold a value I loaded
into it. What happened when further code referenced a property on dequeue?
Undefined, of course. Wonder why this is vues behavior. Caused me.. a headache